+++ src/components/chat/chatsnap/ChatSnap.tsx
@@ -1,5 +1,5 @@
 // React + Web3 Essentials
-import React from 'react';
+import React from "react";
 
 // External Packages
 import styled, { useTheme } from 'styled-components';
@@ -11,21 +11,21 @@ import { useResolveEns } from 'hooks/useResolveEns';
 
 // Internal Configs
 import { appConfig } from 'config';
-import GLOBALS from 'config/Globals';
+import GLOBALS from "config/Globals";
 
 // Interfaces
 export interface ChatSnapMsgI {
-  type: string;
-  message: string;
+  type: string,
+  message: string
 }
 
 interface ChatSnapPropsI {
-  pfp: string;
-  username: string;
-  chatSnapMsg: ChatSnapMsgI;
-  timestamp: number;
-  selected: boolean;
-  onClick?: Function;
+  pfp: string,
+  username: string,
+  chatSnapMsg: ChatSnapMsgI,
+  timestamp: number,
+  selected: boolean,
+  onClick?: Function
 }
 
 // Other Information section
@@ -35,41 +35,37 @@ const ChatSnap = ({ pfp, username, chatSnapMsg, timestamp, selected, onClick }: 
 
   // get ens name
   const ensName = useResolveEns(username);
+  // get reverse name
+ 
 
   // get short username
   const shortUsername = caip10ToWallet(username).slice(0, 8) + '...' + caip10ToWallet(username).slice(-7);
 
   // format message here instead
-  const message =
-    chatSnapMsg.type === 'Text' ? (
-      <SpanV2 color={theme.default.secondaryColor}>
-        {chatSnapMsg.message.length > 25 ? chatSnapMsg.message.slice(0, 25) + '...' : chatSnapMsg.message}
-      </SpanV2>
-    ) : chatSnapMsg.type === 'Image' ? (
-      <SpanV2 color={theme.default.secondaryColor}>
-        <i
-          className="fa fa-picture-o"
-          aria-hidden="true"
-        ></i>{' '}
-        Image
-      </SpanV2>
-    ) : chatSnapMsg.type === 'File' ? (
-      <SpanV2 color={theme.default.secondaryColor}>
-        <i
-          className="fa fa-file"
-          aria-hidden="true"
-        ></i>{' '}
-        File
-      </SpanV2>
-    ) : chatSnapMsg.type === 'GIF' ? (
-      <SpanV2 color={theme.default.secondaryColor}>
-        <i
-          className="fa fa-picture-o"
-          aria-hidden="true"
-        ></i>{' '}
-        GIF
-      </SpanV2>
-    ) : null;
+  const message = chatSnapMsg.type === 'Text' ? (
+    <SpanV2 color={theme.default.secondaryColor}>
+      {chatSnapMsg.message.length > 25 ? (
+        chatSnapMsg.message.slice(0, 25) + '...'
+      ) : (
+        chatSnapMsg.message
+      )}
+    </SpanV2>
+  ) : chatSnapMsg.type === 'Image' ? (
+    <SpanV2 color={theme.default.secondaryColor}>
+      <i className="fa fa-picture-o" aria-hidden="true"></i>
+      {" "}Image
+    </SpanV2>
+  ) : chatSnapMsg.type === 'File' ? (
+    <SpanV2 color={theme.default.secondaryColor}>
+      <i className="fa fa-file" aria-hidden="true"></i>
+      {" "}File
+    </SpanV2>
+  ) : chatSnapMsg.type === 'GIF' ? (
+    <SpanV2 color={theme.default.secondaryColor}>
+      <i className="fa fa-picture-o" aria-hidden="true"></i>
+      {" "}GIF
+    </SpanV2>
+  ) : null;
 
   // get date
   let date = null;
@@ -77,7 +73,7 @@ const ChatSnap = ({ pfp, username, chatSnapMsg, timestamp, selected, onClick }: 
     const time = new Date(timestamp);
     date = time.toLocaleTimeString('en-US').slice(0, -6) + time.toLocaleTimeString('en-US').slice(-2);
   }
-
+  
   // RENDER
   return (
     <ChatSnapContainer
@@ -89,51 +85,41 @@ const ChatSnap = ({ pfp, username, chatSnapMsg, timestamp, selected, onClick }: 
       hover={theme.chat.snapFocusBg}
       hoverBackground={selected ? theme.chat.snapFocusBg : 'transparent'}
     >
-      <ItemVV2
-        width="48px"
-        maxWidth="48px"
-        borderRadius="100%"
-        overflow="hidden"
-        margin="0 5px 0 0"
-      >
-        <ImageV2
-          alt={`Profile pic of ${username}`}
-          src={pfp}
-        />
+      <ItemVV2 width="48px" maxWidth="48px" borderRadius="100%" overflow="hidden" margin="0 5px 0 0">
+        <ImageV2 alt={`Profile pic of ${username}`} src={pfp} />
       </ItemVV2>
 
       <ItemVV2 margin="0 0 0 5px">
-        <ItemHV2
-          alignItems="flex-start"
-          margin="2px 0 2px 0"
-        >
-          <SpanV2
+        
+        <ItemHV2 alignItems="flex-start" margin="2px 0 2px 0">
+          <SpanV2 
             // color={ensName ? '#d53793' : theme.default.color}
             color={theme.default.color}
-            fontWeight={ensName ? '600' : '500'}
+            fontWeight={ensName ? "600" : "500"}
             textAlign="start"
             flex="1"
             fontSize="14px"
           >
-            {ensName && ensName}
-            {!ensName && shortUsername}
+            {ensName && 
+              ensName
+            }
+            {!ensName && 
+              shortUsername
+            }
           </SpanV2>
-          {date && (
-            <SpanV2
+          {date && 
+            <SpanV2 
               color={theme.default.secondaryColor}
               fontWeight="500"
               fontSize="12px"
             >
-              {date}
-            </SpanV2>
-          )}
+            {date}
+          </SpanV2>
+          }
         </ItemHV2>
 
-        <ItemHV2
-          alignItems="flex-end"
-          margin="2px 0 2px 0"
-        >
-          <SpanV2
+        <ItemHV2 alignItems="flex-end" margin="2px 0 2px 0">
+          <SpanV2 
             color={theme.default.secondaryColor}
             flex="1"
             textAlign="start"
@@ -142,13 +128,14 @@ const ChatSnap = ({ pfp, username, chatSnapMsg, timestamp, selected, onClick }: 
             {message}
           </SpanV2>
         </ItemHV2>
+
       </ItemVV2>
     </ChatSnapContainer>
   );
-};
+}
 export default ChatSnap;
 
 const ChatSnapContainer = styled(ButtonV2)`
   flex-direction: row;
   align-self: stretch;
-`;
+`
