+++ src/components/chat/w2wChat/chatBox/ChatBox.tsx
@@ -24,7 +24,7 @@ import * as w2wHelper from 'helpers/w2w/';
 import { generateKeyPair } from 'helpers/w2w/pgp';
 import useToast from 'hooks/useToast';
 import { useResolveEns } from 'hooks/useResolveEns';
-import { AppContext, Context } from 'sections/chat/ChatMainSection';
+import { AppContext, Context } from 'modules/chat/ChatModule';
 import HandwaveIcon from '../../../../assets/chat/handwave.svg';
 import { caip10ToWallet, decryptAndVerifySignature, encryptAndSign, walletToCAIP10 } from '../../../../helpers/w2w';
 import { fetchInbox, fetchIntent, MessageIPFS } from 'helpers/w2w/ipfs';
