+++ src/components/chat/w2wChat/w2wUtils.ts
@@ -1,20 +1,9 @@
-import { Feeds, getFromIPFS, getInbox } from "api"
-import { InboxChat } from "'modules/chat/ChatModule';"
-
-export interface MessageIPFS {
-  fromCAIP10: string
-  toCAIP10: string
-  fromDID: string
-  toDID: string
-  messageType: string
-  messageContent: string
-  signature: string
-  sigType: string
-  link: string | null
-  timestamp?: number
-  encType: string
-  encryptedSecret: string
-}
+import { ConnectedUser, Feeds, getFromIPFS, getInbox, User } from 'api'
+import { DID } from 'dids'
+import { decryptAndVerifySignature } from 'helpers/w2w'
+import { MessageIPFS } from 'helpers/w2w/ipfs'
+import { InboxChat } from 'modules/chat/ChatModule'
+import { intitializeDb } from './w2wIndexeddb'
 
 export const fetchMessagesFromIPFS = async (inbox: Feeds[]): Promise<Feeds[]> => {
   for (const i in inbox) {
@@ -85,4 +74,38 @@ export const fetchIntent = async ({
   }
   intents = await fetchMessagesFromIPFS(intents)
   return intents
+}
+
+export const formatFileSize = (size: number): string => {
+  const i = Math.floor(Math.log(size) / Math.log(1024))
+  return `${(size / Math.pow(1024, i)).toFixed(1)} ${['B', 'KB', 'MB', 'GB', 'TB'][i]}`
+}
+
+export const decryptFeeds = async ({
+  feeds,
+  connectedUser,
+}: {
+  feeds: Feeds[]
+  connectedUser: ConnectedUser
+}): Promise<Feeds[]> => {
+  for (let feed of feeds) {
+    if (feed.msg.encType !== 'PlainText' && feed.msg.encType !== null) {
+      // To do signature verification it depends on who has sent the message
+      let signatureValidationPubliKey: string
+      if (feed.msg.fromCAIP10 === connectedUser.wallets.split(',')[0]) {
+        signatureValidationPubliKey = connectedUser.publicKey
+      } else {
+        signatureValidationPubliKey = feed.publicKey
+      }
+
+      feed.msg.lastMessage = await decryptAndVerifySignature({
+        cipherText: feed.msg.lastMessage,
+        encryptedSecretKey: feed.msg.encryptedSecret,
+        publicKeyArmored: signatureValidationPubliKey,
+        signatureArmored: feed.msg.signature,
+        privateKeyArmored: connectedUser.privateKey
+      })
+    }
+  }
+  return feeds
 }
