+++ src/hooks/useResolveEns.ts
@@ -3,7 +3,7 @@ import { ethers } from 'ethers';
 import { useContext,useEffect, useState } from 'react';
 // Internal Components
 import { caip10ToWallet } from 'helpers/w2w';
-import { AppContext, Context } from 'sections/chat/ChatMainSection';
+import { AppContext, Context } from 'modules/chat/ChatModule';
 
 // Internal Configs
 import { appConfig } from 'config';
